Extract MediaSessionCallback into its own file
